# apps/docs/src/openapi/tradr-api.json
{
...
  "info": {
    "description": "HTTP API for Tradr — the open-source trading journal with an AI advisor.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Tradr API",
    "version": "0.5.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/positions": {
      "post": {
        "description": "Authed. Creates a draft position in one of the user's accounts. When feature gating is enabled and the user is a non-admin Free user, the create is refused with `403 TIER_LIMIT_POSITIONS` at the L2 position cap, or `403 TIER_ACCOUNT_NOT_WRITABLE` when over the L1 account cap and targeting an account other than the writable designation (plan-tiers REQ-6.1/6.6). Admins and gating-off deployments pass through unchanged. Existing-data management (close/fills/edit/delete) is never blocked (REQ-6.5).\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "accountId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "assetType": {
                    "enum": [
                      "stock",
                      "option"
                    ],
                    "type": "string"
                  },
                  "notes": {
                    "nullable": true,
                    "type": "string"
                  },
                  "side": {
                    "enum": [
                      "long",
                      "short"
                    ],
                    "type": "string"
                  },
                  "stopLoss": {
                    "description": "Trade-plan stop-loss price. Positive decimal string with up to 8 decimal places, or null. Used to compute risk/reward.\n",
                    "example": "140.00",
                    "nullable": true,
                    "type": "string"
                  },
                  "symbol": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "targetPrice": {
                    "description": "Trade-plan target price. Positive decimal string with up to 8 decimal places, or null. Used to compute the target R/R.\n",
                    "example": "150.00",
                    "nullable": true,
                    "type": "string"
                  }
                },
                "required": [
                  "accountId",
                  "symbol",
                  "side",
                  "assetType"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created position."
          },
          "400": {
            "description": "Validation error."
          },
          "403": {
            "description": "TIER_LIMIT_POSITIONS (position cap reached) or TIER_ACCOUNT_NOT_WRITABLE (non-writable account while over the account cap)."
          },
          "404": {
            "description": "Account not found (or not owned by the user)."
          }
        },
        "summary": "Create a position (draft).",
        "tags": [
          "Positions"
        ]
      }
    },
    "/api/positions/{id}": {
      "delete": {
        "description": "Authed. Hard-deletes a position in any status (`draft`, `open`, or `closed`) and cascades to its fills. When the position had written a ledger entry via its close, a compensating `position_pnl_reversal` row is posted in the same transaction so the account balance is not left inflated; the ledger stays append-only (`positionId` is set NULL on the affected rows, which net to zero in the balance).\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Position deleted; its fills cascade and any prior close is reversed in the ledger."
          },
          "404": {
            "description": "Position not found (or not owned by the user)."
          }
        },
        "summary": "Delete a position (any status).",
        "tags": [
          "Positions"
        ]
      },
      "put": {
        "description": "Authed. All fields optional. On a `draft` position every field is editable. On `open` and `closed` positions only `notes`, `targetPrice`, and `stopLoss` may change — the trade-plan fields are plan annotations that do not affect fill quantities; any other field on a non-draft position is rejected with 409.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "accountId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "assetType": {
                    "enum": [
                      "stock",
                      "option"
                    ],
                    "type": "string"
                  },
                  "notes": {
                    "nullable": true,
                    "type": "string"
                  },
                  "side": {
                    "enum": [
                      "long",
                      "short"
                    ],
                    "type": "string"
                  },
                  "stopLoss": {
                    "description": "Trade-plan stop-loss price. Positive decimal string with up to 8 decimal places, or null. Editable on any status.\n",
                    "example": "140.00",
                    "nullable": true,
                    "type": "string"
                  },
                  "symbol": {
                    "maxLength": 20,
                    "minLength": 1,
                    "type": "string"
                  },
                  "targetPrice": {
                    "description": "Trade-plan target price. Positive decimal string with up to 8 decimal places, or null. Editable on any status.\n",
                    "example": "150.00",
                    "nullable": true,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated position."
          },
          "400": {
            "description": "Validation error (e.g. an invalid option symbol or plan price)."
          },
          "404": {
            "description": "Position not found (or not owned)",
            "or the target account not found.": null
          },
          "409": {
            "description": "A restricted field (symbol/side/assetType/accountId) was changed on a non-draft position",
            "or an asset-type change was attempted.": null
          }
        },
        "summary": "Update a position.",
        "tags": [
          "Positions"
        ]
      }
    },
...
    "/api/positions/{id}/reopen": {
      "post": {
        "description": "Authed. Transitions a `closed` position back to `open` for intraday re-entries, clearing `closedAt` while preserving `openedAt` (the position's identity day). Allowed only when `openedAt` and the reopen timestamp fall on the same calendar day **in the account's timezone** — a position opened on a previous day must be re-entered as a new position. A supplied `reopenedAt` may not precede `closedAt` nor lie in the future. Reopening posts a compensating reversing ledger row for the prior close so a subsequent re-close does not double-count realized P&L.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reopenedAt": {
                    "description": "Optional ISO-8601 reopen timestamp; defaults to now. Must be on or after the position's close time and not in the future.\n",
                    "format": "date-time",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "The reopened (now open) position."
          },
          "400": {
            "description": "reopenedAt precedes closedAt or is in the future."
          },
          "404": {
            "description": "Position not found (or not owned by the user)."
          },
          "409": {
            "description": "The position is not closed",
            "or its open day (in the account timezone) differs from the reopen day.": null
          }
        },
        "summary": "Reopen a same-day closed position.",
        "tags": [
          "Positions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}